+++ Tests/Source/Rendering/Mesh/TestMeshData.cpp
@@ -3,8 +3,8 @@
 
 TEST(TestMeshData, TestPreAllocate) {
     FTMeshData<FTVertexColor<glm::vec3>> mesh(40);
-    EXPECT_EQ(mesh.getVertices().capacity(), 40);
-    EXPECT_EQ(mesh.getVertexCount(), 0);
+    EXPECT_EQ(mesh.getVertices().capacity(), (size_t)40);
+    EXPECT_EQ(mesh.getVertexCount(), (size_t)0);
 }
 
 TEST(TestMeshData, TestDynamic) {
@@ -14,7 +14,7 @@ TEST(TestMeshData, TestDynamic) {
     for (int i = 0; i < 15; i++) {
         vertices.push_back(FTVertexColor<glm::vec3>(glm::vec3(i, i, i), glm::vec3(0, 0, 0)));
     }
-    EXPECT_EQ(mesh.getVertexCount(), 15);
+    EXPECT_EQ(mesh.getVertexCount(), (size_t)15);
 }
 
 TEST(TestMeshData, TestData) {
@@ -23,9 +23,9 @@ TEST(TestMeshData, TestData) {
     for (int i = 0; i < 10; i++) {
         vertices[i] = FTVertexColor<glm::vec3>(glm::vec3(i, i, i), glm::vec3(0, 0, 0));
     }
-    EXPECT_EQ(mesh.getVertexCount(), 0);
+    EXPECT_EQ(mesh.getVertexCount(), (size_t)0);
     mesh.setVertexCount(10);
-    EXPECT_EQ(mesh.getVertexCount(), 10);
+    EXPECT_EQ(mesh.getVertexCount(), (size_t)10);
 
     EXPECT_THROW(mesh.setVertexCount(15), FTException);
 }
